View карт выносится в отдельный файл
